requirements.txt
pip
requests
fastapi[all]
gunicorn
uvicorn
python-dateutil
pytest
pytest-cov
coveralls